- Created the update notification server - Added the notification model - Created notifications and linked to frontend - Updated Navbar - Refactored code - Created server actions for post created and edited

+++ src/action/session/delete-session.tsx
@@ -4,6 +4,7 @@ import { db } from '@/lib/db';
 import { getSessionById, getSessionUserBySessionUserId } from '@/data/session';
 import { currentUser } from '@/lib/authentication';
 import { getUserById } from '@/data/user';
+import { isAfter } from 'date-fns';
 import { revalidateTag } from 'next/cache';
 
 export const delete_session = async (sessionId: string) => {
@@ -19,6 +20,11 @@ export const delete_session = async (sessionId: string) => {
   if (!session) {
     return { error: 'Session does not exist' };
   }
+  const now = new Date();
+  const isSessionEnded = isAfter(now, session.endDateTime as Date);
+  if (session.users.length > 1 && !isSessionEnded) {
+    return { error: 'Not allowed to delete session' };
+  }
   const sessionParticipant = await getSessionUserBySessionUserId(
     sessionId,
     dbUser.id
@@ -26,11 +32,15 @@ export const delete_session = async (sessionId: string) => {
   if (!sessionParticipant) {
     return { error: 'You are not a member of this session' };
   }
-  await db.sessionParticipant.delete({
-    where: {
-      id: sessionParticipant.id,
-    },
-  });
+  if (sessionParticipant.userRole !== 'ADMIN') {
+    return { error: 'You are not authorised to delete session' };
+  }
+
+  // await db.sessionParticipant.delete({
+  //   where: {
+  //     id: sessionParticipant.id,
+  //   },
+  // });
   await db.session.delete({
     where: {
       id: sessionId,
